feat(abstract-utxo): move transaction types to separate file
Refactor transaction-related types out of abstractUtxoCoin.ts into a new
transaction/types.ts file. This helps organize code better and reduces
coupling between modules.

Move BaseOutput, FixedScriptWalletOutput, Output, BaseParsedTransactionOutputs,
BaseParsedTransaction, and ParsedTransaction to the new file. Also moved
isWalletOutput helper to its relevant module.

Issue: BTC-2732
